fixed progress percent formating status bar is now bigger git optamizer now checks if its actually needed first

+++ GitItGUI.Core/BranchManager.cs
@@ -133,7 +133,7 @@ namespace GitItGUI.Core
 					var options = new CheckoutOptions();
 					options.OnCheckoutProgress = new LibGit2Sharp.Handlers.CheckoutProgressHandler(delegate(string path, int completedSteps, int totalSteps)
 					{
-						if (statusCallback != null) statusCallback(string.Format("Checking out: {0}%", ((completedSteps / (decimal)(totalSteps+1)) * 100)));
+						if (statusCallback != null) statusCallback(string.Format("Checking out: {0}%", (int)((completedSteps / (decimal)(totalSteps+1)) * 100)));
 					});
 
 					Filters.GitLFS.statusCallback = statusCallback;
@@ -178,7 +178,7 @@ namespace GitItGUI.Core
 				var options = new MergeOptions();
 				options.OnCheckoutProgress = new LibGit2Sharp.Handlers.CheckoutProgressHandler(delegate(string path, int completedSteps, int totalSteps)
 				{
-					if (statusCallback != null) statusCallback(string.Format("Merging: {0}%", ((completedSteps / (decimal)(totalSteps+1)) * 100)));
+					if (statusCallback != null) statusCallback(string.Format("Merging: {0}%", (int)((completedSteps / (decimal)(totalSteps+1)) * 100)));
 				});
 
 				var srcBround = RepoManager.repo.Branches[srcBranch.fullName];
